Move NotAuthorised to components and add Twilio dependency
NotAuthorised component is now under the components directory for better organization of the project files. Additionally, the Twilio SDK dependency was added to the project's backend to enable future use of Twilio's communication APIs.
