+++ backend/pom.xml
@@ -59,6 +59,11 @@
 				<version>RELEASE</version>
 				<scope>compile</scope>
 			</dependency>
+			<dependency>
+				<groupId>com.twilio.sdk</groupId>
+				<artifactId>twilio</artifactId>
+				<version>8.19.0</version>
+			</dependency>
             <dependency>
                 <groupId>org.jetbrains</groupId>
                 <artifactId>annotations</artifactId>
